Fixed build on S60 and WinCE
Since qreal is a float on these targets, it was having trouble picking
the right version of qMax.

+++ src/gui/painting/qpainter.cpp
@@ -5918,7 +5918,7 @@ static QPixmap generateWavyPixmap(qreal maxRadius, const QPen &pen)
     if (QPixmapCache::find(key, pixmap))
         return pixmap;
 
-    const qreal halfPeriod = qMax(qreal(2), radiusBase * 1.61803399); // the golden ratio
+    const qreal halfPeriod = qMax(qreal(2), qreal(radiusBase * 1.61803399)); // the golden ratio
     const int width = qCeil(100 / (2 * halfPeriod)) * (2 * halfPeriod);
     const int radius = qFloor(radiusBase);
